LLM Projects/Web-Scraper/requirements.txt
 google-generativeai
 streamlit
 python-dotenv
 requests
 beautifulsoup4